% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{checkCohortReference}
\alias{checkCohortReference}
\title{Load Cohort Definitions From A Study Package}
\usage{
checkCohortReference(cohortReference, errorMessage = NULL)
}
\description{
Checks cohort references 
Checks cohort references
}
