% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFunctions.R
\name{divide_csv_output_by_DEC}
\alias{divide_csv_output_by_DEC}
\title{Divide output csv-files into smaller csv-files by DEC groups}
\usage{
divide_csv_output_by_DEC(folder_path = "export", cohortids_per_dec = NULL)
}
\arguments{
\item{folder_path}{The path to where the csv-files to slice-and-dice lives.}

\item{cohortids_per_dec}{A data frame containing two integer columns cohort_id and dec}
}
\value{
Nothing, but stores as many folders as there are DECs on the folderPath.
}
\description{
Reads in all output csv files, and creates separate zip-files for each DEC.
}
