% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronograph-SARAV01.R, R/Chronograph.R
\name{executeChronograph}
\alias{executeChronograph}
\title{Get the data for chronographs from the server.}
\usage{
executeChronograph(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = c(),
  targetCohortId = c(),
  comparatorCohortId = c(),
  outcomeCohortId = c(),
  cdmVersion = "5",
  shrinkage = 0.5,
  icPercentile = 0.025,
  project_root_path = project_root_path,
  verbose = FALSE
)

executeChronograph(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = c(),
  targetCohortId = c(),
  comparatorCohortId = c(),
  outcomeCohortId = c(),
  cdmVersion = "5",
  shrinkage = 0.5,
  icPercentile = 0.025,
  project_root_path = project_root_path,
  verbose = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{A string with the name of database schema that contains OMOP CDM and vocabulary.
For sql server, append the schema name to the database name as "[database_name].[schema_name]" 
and use as cdmDatabaseSchema.}

\item{oracleTempSchema}{For Oracle only: A string with the name of the database schema where you want all
temporary tables to be managed. Requires create/insert permissions
to this database.}

\item{cohortDatabaseSchema}{A string with the name of the database schema that is the location where the
cohort data is available.  For sql server, append the schema name to 
the database name as "[database_name].[schema_name]" and use as 
cohortDatabaseSchema. If not specified, cdmDatabaseSchema will be used.}

\item{cohortTable}{The tablename that contains the three required cohorts.}

\item{targetCohortId}{A numeric variable containing the CohortID of the target cohort.}

\item{comparatorCohortId}{A numeric variable containing the CohortID of the comparator cohort.}

\item{outcomeCohortId}{A numeric variable containing the CohortID of the outcome cohort.}

\item{cdmVersion}{Define the OMOP CDM version used: currently supports "5".}

\item{shrinkage}{Shrinkage used in IRR calculations, required >0 to deal with 0 case
counts, but larger number means more shrinkage. default is 0.5}

\item{icPercentile}{The lower bound of the credibility interval for the IC values
(IClow). default is 0.025,}

\item{project_root_path}{The path to the project root.}

\item{verbose}{Should informative messages be printed?}
}
\value{
A data frame with observed and expected outcome counts in periods relative to the exposure
initiation date, for each outcome and exposure.

A data frame with observed and expected outcome counts in periods relative to the exposure
initiation date, for each outcome and exposure.
}
\description{
Get the data for creating chronographs from the server.

Get the data for creating chronographs from the server.
}
