% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronograph.R
\name{executeChronographWrapper}
\alias{executeChronographWrapper}
\title{Wrapper to execute chronograph scripts}
\usage{
executeChronographWrapper(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = paste0("cohorts_", Sys.getenv("USERNAME"), "_CD"),
  decList = "../inst/settings/DecList.csv",
  storePath = "export",
  storeFileName = "chronograph_data.csv",
  databaseId = cdmDatabaseSchema,
  project_root_path = NULL,
  create_pngs_in_extras_folder = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{A string with the name of database schema that contains OMOP CDM and vocabulary.
For sql server, append the schema name to the database name as "[database_name].[schema_name]" 
and use as cdmDatabaseSchema.}

\item{oracleTempSchema}{For Oracle only: A string with the name of the database schema where you want all
temporary tables to be managed. Requires create/insert permissions
to this database.}

\item{cohortDatabaseSchema}{A string with the name of the database schema that is the location where the
cohort data is available.  For sql server, append the schema name to 
the database name as "[database_name].[schema_name]" and use as 
cohortDatabaseSchema. If not specified, cdmDatabaseSchema will be used.}

\item{decList}{A path to the Drug-event-combination list.}

\item{storePath}{If not specified defaults to "export"}

\item{storeFileName}{If not specified defaults to chronograph_data.csv}

\item{project_root_path}{The path to the project root}

\item{create_pngs_in_extras_folder}{A quick-fix - save the chronograph plots in the extras folder?}
}
\value{
Nothing, but stores a csv with all chronograph data.
}
\description{
A for-loop wrapper to execute chronograph script on a list of drug-event-combinations (DECs).
}
