% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{getInclusionStats}
\alias{getInclusionStats}
\title{Get Inclusion Stats}
\usage{
getInclusionStats(
  connection,
  exportFolder,
  databaseId,
  cohortDefinitionSet,
  cohortDatabaseSchema,
  cohortTableNames,
  incremental,
  instantiatedCohorts,
  inclusionStatisticsFolder,
  minCellCount,
  recordKeepingFile
)
}
\description{
Collects statistics on inclusion.
}
