% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getResultsDataModelSpecifications_mod}
\alias{getResultsDataModelSpecifications_mod}
\title{Get specifications for Cohort Diagnostics results data model}
\usage{
getResultsDataModelSpecifications_mod()
}
\value{
A tibble data frame object with specifications
}
\description{
Get specifications for Cohort Diagnostics results data model
}
