% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{loadCohortsFromPackage}
\alias{loadCohortsFromPackage}
\title{Load Cohort Definitions From A Study Package}
\usage{
loadCohortsFromPackage(
  packageName,
  cohortToCreateFile = "settings/cohortsToCreate.csv",
  cohortIds = NULL,
  errorMessage = NULL
)
}
\arguments{
\item{errorMessage}{checkmate assert collection, used internally for error checks}
}
\description{
Load cohort references for usage in executeDiagnostics.
}
