% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomFunctions.R
\name{loadCsvIntoPostGreSQL}
\alias{loadCsvIntoPostGreSQL}
\title{Title}
\usage{
loadCsvIntoPostGreSQL(
  connectionDetails = postGresConnectionDetails,
  project_root_path = "path_to_project_root",
  export_folder = exportFolder_after_slash,
  schema = schema_to_insert_into,
  dropTableIfExists = TRUE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{project_root_path}{A string containing the path to the root of the Rstudio project.}

\item{schema}{A string of the schema where the CohortDiagnostics-backend data is to be placed.}
}
\value{
Returns nothing.
}
\description{
This function loads the csvs created by CohortDiagnostics into an (initiated manually outside of this function) PostGreSQL-database. This serves as a backend to the shiny app which hosts CohortDiagnostics. The function takes all csv-files in the export folder and tries to load them into the specified schema of the PostGreSQL database.
}
