% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronograph-SARAV01.R, R/Chronograph.R
\name{plotChronograph}
\alias{plotChronograph}
\title{Plot a chronograph}
\usage{
plotChronograph(
  data = result_df,
  targetCohortId,
  comparatorCohortId,
  outcomeCohortId,
  title = "",
  fileName = NULL
)

plotChronograph(
  data = result_df,
  targetCohortId,
  comparatorCohortId,
  outcomeCohortId,
  title = "",
  fileName = NULL
)
}
\arguments{
\item{data}{Data as generated using the \code{getChronographData} function.}

\item{targetCohortId}{The cohortID for the target cohort}

\item{comparatorCohortId}{The cohortID for the comparator cohort}

\item{outcomeCohortId}{The cohortID for the outcome}

\item{title}{The title to show above the plot.}

\item{fileName}{Name of the file where the plot should be saved, with file format ending, e.g. 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\description{
Creates a plot showing the observed and expected number of outcomes for each month in the specified time interval 
before and after initiation of the exposure, as well as the IC. See references for further details.

Creates a plot showing the observed and expected number of outcomes for each month in the specified time interval 
before and after initiation of the exposure, as well as the IC. See references for further details.
}
\references{
Noren GN, Hopstadius J, Bate A, Star K, Edwards R, Temporal pattern discovery in longitudinal
electronic patient records, Data Mining and Knowledge Discovery, May 2010, Volume 20, Issue 3, pp
361-387.

Noren GN, Hopstadius J, Bate A, Star K, Edwards R, Temporal pattern discovery in longitudinal
electronic patient records, Data Mining and Knowledge Discovery, May 2010, Volume 20, Issue 3, pp
361-387.
}
